requirements.txt
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
psycopg2-binary==2.8.5
uuid==1.30